requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
dateparser==1.1.2
fonttools==4.38.0
frozenlist==1.3.1
google-auth==2.13.0
google-auth-oauthlib==0.5.3
gspread==5.6.0
idna==3.3
kiwisolver==1.4.4
matplotlib==3.6.1
multidict==6.0.2
numpy==1.23.2
oauthlib==3.2.2
packaging==21.3
pandas==1.4.4
Pillow==9.3.0
plotly==5.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==4.2.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tenacity==8.1.0
tzdata==2022.6
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.12
websockets==10.4
yarl==1.8.1
